requirements.txt
-i https://pypi.org/simple
beautifulsoup4==4.9.0
bs4==0.0.1
cachetools==4.1.0
certifi==2020.4.5.1
chardet==3.0.4
click==7.1.1
flask==1.1.2
google-api-core[grpc]==1.16.0
google-auth==1.13.1
google-cloud-core==1.3.0
google-cloud-firestore==1.6.2
googleapis-common-protos==1.51.0
grpcio==1.28.1
gunicorn==20.0.4
idna==2.9
itsdangerous==1.1.0
jinja2==2.11.2
markupsafe==1.1.1
protobuf==3.11.3
pyasn1-modules==0.2.8
pyasn1==0.4.8
python-dotenv==0.12.0
pytz==2019.3
requests==2.23.0
rsa==4.0
six==1.14.0
soupsieve==2.0
urllib3==1.25.8
werkzeug==1.0.1
